Expose HTML tabs files in a specific /tabs URI
Wok provides a specific API to load tabs content (/tabs URI). It is designed to
identify when a tab is requested to properly setup the 'lastPage' cookie.
That way, the last tab visited by user can be remembered.

This patch depends on Wok patch:
- Bug fix: Make /tabs URI generic to be used by any plugin

To test it, you can login into Wok, access a different tab than the default,
logout, login again and you will be redirected to your last selection.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>