Create read-only view for Dashboard tab
While accessing Ginger + GingerBase with a normal, no tab is displayed
because all of them were restricted to admin users.
As Dashboard tab only contains informative data, it was set as read-only.
So a normal user can see the tab content but can not perform any action
(reboot/shutdown).

In future, it would be better to have those actions (reboot/shutdown)
moved to Administration tab (part of Ginger).

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>