Fixing broken packaging due to wrong variable
The patch 'Expose HTML tabs files in a specific /tabs URI' broke
the RPM build (and probably the .deb build as well) because it
created a variable

tabshtmldir = $(datadir)/kimchi/ui/pages/tabs

in the ui/pages/tabs/Makefile.am file. Probably because the patch
came from something that was done in Kimchi. Fixing the variable
to

tabshtmldir = $(datadir)/gingerbase/ui/pages/tabs

solved the issue.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>