Fixing RPM packaging of tabs
This is a follow-up of the previous fix after testing the
resulting RPM in a Fedora 23.

It was necessary to add the tabs path in the %files section
of contrib/gingerbase.spec.fedora.in . Otherwise the UI
fails to find the tabs after the installation by RPM.

Due to that it was also needed to change (again) the
'tabshtmldir' variable of ui/pages/tabs/Makefile.am to:

tabshtmldir = $(datadir)/wok/plugins/gingerbase/ui/pages/tabs

which shouldn't come as a surprise because this is the same
pattern/value that worked out in Ginger.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>