Issue #12: fixing _get_cpus() of HostModel
This code was invoking psutil.NUM_CPUS before checking
if the attribute exists. In case it doesn't (seems to
be the case in Fedora 23), this was causing an uncaught
exception and plugins/gingerbase/host API returning empty.
The UI is not ready to deal with this response and, in
this case, the Dashboard doesn't load.

If we have no means of getting the 'online_cpus' attribute
the values of 'online_cpus' and 'offline_cpus' now defaults
to "unknown".

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>