Github #18: lscpu output varies with system language
At this moment we're doing the parsing of lscpu based
on strings such as "Thread(s)" and "Core(s)" and so on. These
strings are translated to the current system language and
breaks this parsing (in the example mentioned in the Github
issue, the line "Core(s) per socket" in a French system
becomes "Coeur(s) par socket").

Forcing the output of lscpu to be always in the default
language (English) solves this issue.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>