Issue #19: Fix HTTP return code of host updated.
A POST request to /plugins/gingerbase/swupdate is returning "500 Internal Server
Error" when system doesn't have packages to update. However, it's a successful
operation even with no packages to update in the system.

This patch modifies the return of the request to be a dictionary with the
message "No packages marked for update".

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>