Fix debugreport rename
There is bug in debugreport rename function, where it is not possible to
rename a file that shares any substring in the path. Function is going
to replace all substrings in the path and will try to create the file
with the completely wrong path.
This patch fixed the problem

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>