netinfo.py: merging Ginger and Kimchi versions
Both plug-ins are mantaining their own version of this helper,
sharing a good amount of code between them but also with
exclusive methods and features. This single file aims to
merge both versions while trying to minimize the
changes required in Ginger and Kimchi.

A breakdown of how the versions were merged follows. If
a specific function is not mentioned it's because the
code was exactly similar between Ginger and Kimchi.

- operstate and link_detected: both plug-ins interpret
them in a different manner - Ginger users operstate in
both methods, Kimchi uses carrier. None of them were
right - while link_detected is something to read from
the carrier, the iface_status can't reliably be
retrieved from the carrier too, but from operstate.
The decision made is to use the operstate implementation
from Ginger and link_detected from Kimchi.

- get_interface_info: both plug-ins added new parameters
to be returned. This version merges everything:
  * type, ipaddr, netmask, status: common attributes that
remained as is.
  * Ginger uses 'device', Kimchi uses 'name' as the interface
identifier. To avoid breaking APIs both attributes are
now returned.
  * 'link_detected': Kimchi only attribute. Added.
  * 'macaddr' and 'module': Ginger only attributes. Added.

- get_interface_kernel_module, macaddr: Ginger exclusive
functions that were added as is.

- ovs_bridges, is_ovs_bridge, ovs_bridge_ports: Kimchi
exclusive functions that were added as is.

- is_wlan, is_nic, is_bonding, is_vlan, is_bridge,
get_bridge_port_device, is_bare_nic: both plug-ins immplement
those the same way, but Ginger takes encoding into consideration.
Ginger version were kept in this merge.

- bridges, ports, is_brport: similar implementations but
Kimchi versions were kept because they took ovs_bridges
into consideration in the logic and were kept.

- all_interfaces: Kimchi version kept. Ginger version
was simply a call to ethtool.get_devices()

- get_interface_type: Kimchi version kept. Ginger will
need to adjust the return values to the new strings.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>