Remove cherrypy root configuration from gingerbase.conf file
Wok is loading the root configuration for every plugin, so there is no
need to keep it in gingerbase.conf file.

Also the Ginger Base configuration file should only contain data that can be
edited by user, which is not the case of the cherrypy configuration.

This patch also updated the test case as the session timeout is now
handled exclusively by Wok.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>