Do not use default value when declare a function
For mutable object as a default parameter in function and
method-declarations, evaluation and creation takes place at
exactly the same moment. The python-parser reads the
function-head and evaluates it at the same moment.

Most beginners assume that a new object is created at every call, but
that's not correct. One object is created at the moment of declaration
and not on demand when you are calling the method.

For imutable objects that's not a problem, because even if all calls
share the same object, it's imutable and therefore it's properties
remain the same.

As a convention, the None object must be used for defaults to indicate
the use of a default initialization, which now can take place in the
function-body, which naturally is evaluated at call-time.

Signed-off-by: Ramon Medeiros <ramonn@linux.vnet.ibm.com>