swupdate: change lock mechanism and wait for packager to be ready
This patch changes swupdate module as follows:

- do not use gingerbaseLock (Lock based) but a new lock
called swupdateLock which is RLock based (only the thread
that acquired the lock can release it).

- lock/unlock is now made in the SoftwareUpdate class. This is
a Gingerbase control flow that can be done in this level instead
of relying on the specific packager implementation to hold/release
locks.

- a new generic class called GenericUpdate was created as a way to
embed a method 'wait_pkg_manager_available' to all packager
implementations. This method is called in each method of each
packager to ensure that the process will not be blocked by
the external package utility.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>