Make HostStatsHistory optional.
This patch adds the option to turn off the host statistics history as a way to
reduce the CPU consumption of wokd when server is in idle mode.

It adds the configuration option 'hoststats_on' in gingerbase.conf, to turn on
or turn off the thread responsible to collect host statistics and store the
last 60 values. The default value is 'True', that means the collect and store
will run. Changing the value to 'False', means the collect of host statistics
will run only when requested (GET /plugins/gingerbase/host/stats) and store will
not exists.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>