test_rest.py: changing generated reports name
'report1' and 'report2' are common names to be used by
an actual debugreport created by the user. In this case,
test_rest would fail because the report already exists.

This patch changes the names to 'test_rest_report1' and
'test_rest_report2' inside test_rest, making this error
scenario less likely to occur.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>