Bug fix #111: Do not return package dependencies on package lookup()
This patch also updates packagesupdate Collection to do not call
lookup() for each resource. That way, we can run one command to get all
the packages information. It will improve performance and make it
faster.

So
- GET /packagesupdate will return all information related to the
  packages marked to be updated without calling lookup() for each
  resource

- GET /packagesupdate/*name* will call lookup() for the given resource

- GET /packagesupdate/*name*/deps will return the package dependencies
  that need to be updated when updating package *name*

Tested on: CentOS 7, Ubuntu 16.04, Fedora 24 and OpenSUSE 42.1 with more
than 100 packages marked to be updated.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>