Bug fix: Merge repo data with the user changes to update repo information
When updating a repository, user may pass only the value he/she wants to
update and not all them. So merge the repository information with the
user changes to get the final repository information result.

The issue was identified while opening and save the edit repository dialog
without any changes for a YUM repository with metalink information (as
the current UI does not show metalink info and then that info will not
be pass to backend).

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>